Support text/plain target in place of UTF8_STRING
If UTF8_STRING is requested but the clipboard contents has a target of
type "text/plain" it is currently ignored. This adds an exception that
allows text/plain targets to be pasted, but still needs much improvement
to support other valid targets such as STRING, TEXT, text/html etc.

In an ideal world, we should be able to autodetect from/to targets so
that regardless of what kind of string-like target is requested, it'll
always be converted when there's a single string-like target available.